Default MAC address should display "Not Available" in network details
For saved networks that do not have a MAC address available in their
WifiConfiguration such as Passpoint networks, the MAC address will
be the default value of 20:00:00:00:00:00. In this case, display
"Not Available" as the summary of the MAC address preference.

Bug: 137282522
Test: atest WifiDetailPreferenceControllerTest &&
manual - open saved passpoint network details while disconnected
but in range and observe MAC address preference.

Change-Id: Ie9010b0430726de8105852ae2afe10047dcc3676

Former-commit-id: f73d65c940e97b96dd8c572dc63e9df02a834b5d