Don't allow Settings to change its own standby bucket.
When on the "Standby apps" developer options screen, the Settings app
will be greyed out, indicating that its standby bucket cannot be
changed.

Manual Test 1: Go to the "Standby apps" screen in developer options and
observe that the Settings app is greyed out.
Manual Test 2: Turn off "Adaptive Battery" in Settings and then go to
the "Standby apps" screen in developer options - observe that the
Settings app is greyed out.

Bug: 132922935
Test: manual (listed above)
Change-Id: Idb29e6105024feacc270867e46156fc6a0396b01

Former-commit-id: 778d38a551e40ed9d3fe391c5b76bf00a0bf5825