Remove filtering from "recent sent" notifs preference
The 'has a launch intent' filter was inherited from 'recent apps'
preference but an app doesn't need a launch intent in order to
send notifications. Also, the next screen didn't apply the filter
so the data wouldn't always match up.

Also make the preferencecontroller more robust to failure.

Test: manual, trigger a notification from a non-launchable app
and makes sure it appears in the list
Fixes: 141380329
Fixes: 142956641

Change-Id: Icf3abead82c572bbffe6e06ecf51a8e02c11c982

Former-commit-id: f356ced8d7aee75ee2c18debcaf3d87b73e45af6