[WPA3] Hide connect button for cloned networks not in range
Cloned networks are WPA2 or Open networks that have a WPA3 or OWE
cloned networks with the same SSID and credentials. When an AP
in transition mode is in the area, the scan results will only
report the higher security level. In order to prevent a case
where the user forcely connects to the AP with the lower security,
disable this option for those networks. Otherwise, it will look
that Wi-Fi is connected, but Wi-Fi picker will not show a connected
network.

Bug: 143843364
Test: Manual test
Change-Id: Id002b109bd056105ec8335acd298fa3b314c73f3
Merged-In: I7244c41d3bf12ba573c3cfca077d0ac2cfdcf5ed

Former-commit-id: 36b3d9b8cf6feb0fbfcb96bf9e41da8eadd9c181