Add activity for MANAGE_APP_OVERLAY_PERMISSION intent in Settings
In order to support system use-cases, such as when the user wants do
revoke SYSTEM_ALERT_WINDOW (SAW) permission via notification, we are
introducing a new intent, MANAGE_APP_OVERLAY_PERMISSION (notice the
"APP_"), that directs the user to the app-specific screen to toggle SAW
permission.

This activity can only be started if the caller has
INTERNAL_SYSTEM_WINDOW permission, which sysui and the system have.

Test: atest TestManageOverlayPermissionIntents
      (cts in one of the topic CLs)
Test: Let an app draw on top of other apps, via its notification go to
      settings, verify that the screen for controlling such capability
      for the app is dispayed.
Bug: 135920175

Change-Id: I2f7e02a1242af7a22098359aeb230888fb8d4f3a

Former-commit-id: 885403133f3fdb72fab12adf2279931944d4b5e8