Make all TogglePreferenceController slice-able.
All the descendants of TogglePreferenceController are now slice-able by
default, where every controller can still disable it by overriding this
value. Though all of them are now slice-able, doesn't necessarily means
they are exposed publicly as we set the default value of isPublicSlice()
to false. For those that want to be public ones will have to override
this value.

Bug: 141088937
Test: robotests
Change-Id: I2ea529c11f0177cd6c55754b632cd1fea1dcc48f

Former-commit-id: 617458b07c91d59c24e73c7c413c58467e437c7e