Catch crash in mobile network ProgressDialog
When enabling "Automatically select network" in "Settings -> Network &
internet -> Mobile network -> Advanced" and rotating the device while
the dialog is showing there will be an IllegalArgumentException since
the decorView isn't attached to the window when dismissing the dialog.
Fixed by simply catching the exception since the dialog will be gone
anyway.

Bug: 146057372
Test: manual (see comment above)
Change-Id: I15f1d6f505fdf33964ce973ece31d6aa0d8f3909

Former-commit-id: 9f74807e89063fdb5cc0dcae6e8dc45b16d4bf02