Set ClearDefaultsPreference as unselectable for talkback
ClearDefaultsPreference is a customized preference which contains a
TextView and a Button. In reality, the button is the only thing that the user can interact with. It's a no-op when the user clicks on the whole preference view.
 
Mark the ClearDefaultsPreference as unselectable can fix the talkback
issue for the user clicking on the preference view outside of the button.
 
In the meantime, the button of ClearDefaultsPreference has not applied the button style of Settings yet. Also fixed it by using the ActionPrimaryButton style of Settings.


Fixes: 146471192
Test: visual and robotests
Change-Id: Ife215d0901bd9fc3055495b0495ba175ed0dfa26
Merged-In: Ife215d0901bd9fc3055495b0495ba175ed0dfa26
(cherry picked from commit 22238b756c5b0c795e26616a5a687564afc525c2 [formerly b9bcbddc4658636df1b3571747ed886b0240e0fa])

Former-commit-id: 5f503b87eac6f9fccf7bff93055e9d2a3b7c94c0