Disable 5G Scanning in NetworkScanHelper
5G has two operating modes: stand-alone and non-standalone.
The vast majority of support is currently for non-standalone
mode. Because Android currently cannot tell the difference
between the supported modes in the modem, and because a device
that supports non-standalone mode shoudn't scan for 5G during a
PLMN selection procedure (it can't camp on 5G, hence non-
standalone mode), disable 5G scans by default and revisit once
modem support exists to distinguish between the two modes. This
avoids alternatives such as allowing 5G to be silently removed
in the RIL or for a 5G scan to be performed which doesn't provide
the user any meaningful options.

A separate bug has been filed (b/151119451) to revisit this logic
once the modem supports interrogation of its 5G capability.

Bug: 149528140
Test: compilation
Change-Id: I4e19b43e6529b5c673bb54c6f8f94c44a7c5bee5

Former-commit-id: adbc1ac905a8b1e2e70070856b2adda3e183889c