Add a new developer options screen for shared data.
Add a new Storage section in the developer options menu which has a
new Shared Data preference. This preference screen shows all shared data
blobs on the device. There is also a new screen for each data blob which
shows all of the packages which currently have a leases on it. This
screen also has a button to delete the shared data blob.

Bug: 150626561
Test: make RunSettingsRoboTests ROBOTEST_FILTER=SharedDataPreferenceControllerTest
Test: manual (visual)
Change-Id: Id84a33dc7eeac493b2f81d3996ad24ee70557a07

Former-commit-id: 7efe9315bacd3192614fc8a1cdf8207d313f568e