The text on toast should be changed when remains 1 time for SIM lock input
The behavior of getQuantityString() is different according to language, it gets both "other" and "one" string if English is system language. But it always gets "other" string if non-English, then the "other" string of wrong_pin_code is shown even if remains 1 time of SIM lock input.

In order to change the string in any language, we add new string for remains 1 time.

Test: Rebuild rom
Fix: 71783850
Merged-In: If4b53bbbc6b51582bc0837abf4497d9a75ebbaad
Change-Id: I1c069333b19399b727e10ed4a6c9483cc1bc44a6

Former-commit-id: d042a9f2d3ecd6c2e0867b59f29416e76d2e2b57