The text on toast should be changed when remains 1 time for SIM lock input
This cl was cherry-picked from aosp/585306

The behavior of getQuantityString() is different according to language, it gets
both "other" and "one" string if English is system language. But it always gets
"other" string if non-English, then the "other" string of wrong_pin_code is
shown even if remains 1 time of SIM lock input.

In order to change the string in any language, we add new string for
remains 1 time.

Test: Rebuild rom
Fix: 71783850
Change-Id: If4b53bbbc6b51582bc0837abf4497d9a75ebbaad

Former-commit-id: 9c69ded434ffcc182b6cf98c1f1f37b6539f8c34