Fix Talkback annoucement for Media Output panel.
The sound panel and media panel are using the same window, so when it
transit from the sound panel to media panel, no window-change event will
be sent. Therefore, Talkback won't be able to re-focus the panel header
which lead to the lack of the announcement.

Fix this by adding setAccessibilityPanelTitle() to trigger Talkback.

Fixes: 140459762
Test: rebuild and test talkback on the media output panel.
Change-Id: I1442434315374dbbfe1d47ca8638180577c93168

Former-commit-id: 350affcdfac507587118fc31338e93c636c9f403